cleaner organization of tests
